Define LCS as PRIMARY when no IP address specified
After spending some time fighting with the LCS drive, I finally discovered
that, although all parameters are optional, no incoming frames are passed
to the guest if there are no parameters.

If an IP address is specified as a parameter, then it will accept only
frames for that address. To get it to accept other frames, it is necessary
to use an OAT file and define a device with PRI.

Defining an IP addres for the device creates a ROUTE entry, which is not
necessary when using ethernet bridging.

I think the expectation should be that, if no IP address is specified, the
LCS device should accept all frames. My suggesting would be to define the
LCS device as PRIMARY when no IP address is specified, otherwise keep the
existing behaviour.

I made the change against the 3.07 stable code and it works very well,
allowing for simple LCS definition as, e.g.,
 E20.2 LCS

Brian Kennelly