Fix w32util clock_gettime HPC 64-bit integer overflow bug.
The integer arithmetic calculation to convert elapsed HPC
ticks to nanoseconds was multiplying by 1,000,000,000 and
then dividing by HPC ticks per second, but if the elapsed
HPC ticks was sufficiently large (as it is on some hardware
running older versions of Windows) the multiplying by one
billion would overflow the 64-bit integer variable leading
to an elapsed nanoseconds value of zero causing the emulated
TOD clock to stall bringing the guest and sometimes Hercules
itself to a screeching halt.

This fix addresses the problem by range checking the values
used in every calculation to ensure the variables used will
never overflow. As an unavoidable side effect of the change
the resolution of the emulated TOD clock on some systems
running older versions of Windows will be reduced from
nanosecond accuracy to, typically, only 10 nanosecond
accuracy. This is not seen as a problem since the existing
TOD/ETOD clock handling already ensures clock values are
always unique ever increasing values.