Interval timer 2012-04-11 fix completion: remove circumvention code
The fix made in April 2012 to interval timer processing
(git ref: 1e1fa932104292b8d624710e1eb5deadb75fdc9c) was
intended to correct a problem with spurious timer pops
during IPL that the old OS standalone utilities couldn't
handle. It did that, but a piece of an old circumvention
for an interval timer problem got left in by mistake. The
now-extraneous circumvention, combined with the April fix,
caused the location 80 value fetched by the OS external
interrupt handler to get reset to a positive value before
OS was able to look at it, which resulted in OS ignoring
valid timer pops. This fix removes the extraneous code,
and OS now recognizes that a timer pop has occurred.